fix: Fixing new version dialog always popping up
This was due to the Rider code clean up tools removing the return. It
was ultimately my fault for not adding the return back in when I was
reviewing the files changes but realistically I'll keep making that
manual mistake if the manual keeps existing. This should keep it in as
the code will be ignored when the code clean up is run.

closes #23