feat: Adding slider for playback speed based on tests
Tested some settings and examples and it looks like the tempo is the
best way of fast forwarding. Following the rule that tempo +50 == 2x
playback speed.

You can fast forward with rate but it makes you sound like the
chipmunks. You can kind of fix it at lower rates by lowering the pitch
but even at 2x speeds its an obvious distortion pitch aside.

You can also fast forward with beats per minute but it almost seems time
dependant on the length of the TTS and I didn't want to research an
equation. The effect on TTS, when set close to correctly, sounds no
different to me than changing the tempo.

closes #69