feat: switch off using the clipboard for login
Trying something new with the API/deployment environment, WEB SOCKETS!
This really didn't need a web socket it could have been a simple short
poll but you gotta try stuff and have fun.

So the workflow is, we use the ws as a pub/sub. When we open the ws
connection against the API we give it the identifier for our
subscription. Then we launch the web browser in a separate process
telling twitch to authenticate with the user and pass us the code for an
oauth token. Once the twitch api, using the redirect url, calls our API,
the API will publish the oauth credentials to the correct ws.

Again, kinda overkill for what we're trying to accomplish but what is
life without a little fun.